Refactor feature methods into macros
Also adds a method to expect guest remove button to be/not to be present.
